# bug: [fix performance test not working with multiple packages] (FF-3110)