--- a/Makefile
+++ b/Makefile
@@ -29,12 +29,9 @@ lint:
 
 ## profile-memory - Run test and generate memory profile
 profile-memory: test-data
-	@cd eppoclient && \
-	{ \
-	echo "Using OUTFILE_SUFFIX: $$OUTFILE_SUFFIX"; \
-	go test -run Test_e2e -memprofile ../memprofile$$OUTFILE_SUFFIX.out ./...; \
-	go tool pprof -text -nodecount=50 ../memprofile$$OUTFILE_SUFFIX.out > ../memprofile$$OUTFILE_SUFFIX.text; \
-	}
+	@echo "Using OUTFILE_SUFFIX: $(OUTFILE_SUFFIX)"
+	go test -v ./... -run=Test_e2e -memprofile=memprofile$(OUTFILE_SUFFIX).out
+	go tool pprof -text memprofile$(OUTFILE_SUFFIX).out > memprofile$(OUTFILE_SUFFIX).text
 
 ## profile-memory-compare - Compare two memory profiles
 ## example: make profile-memory-compare BASE_FILE=memprofile1.out FEAT_FILE=memprofile2.out