--- a/doc.go
+++ b/doc.go
@@ -1,1 +1,0 @@
-package sdk