On each PR commit we run a github action to compare memory allocations in the e2e test suite between the commit in the PR and the latest main.

Cannot run the memory profiler with multiple packages:

```
cannot use -memprofile flag with multiple packages
```

A nested package was added in https://github.com/Eppo-exp/golang-sdk/commit/1e1323b342f91f620e198abdc07ed6bbb61710cb

## Changes

* Removes the `applicationlogger` package, flattens into main `eppoclient` package.
* Runs the test 5x instead of once to get more memory allocations and further separate the comparison
* housekeeping like upgrading github action versions

👉 It is expected that the `memory profile comparison` test still fails since part of its job is to checkout code from main branch, which is of course currently broken.

https://github.com/Eppo-exp/golang-sdk/actions/runs/10746934858/job/29808652096?pr=68

I switched the order of the execution to first perform the memory profiling against the feature branch (this one) and it passed, so I am confident that after this PR is merged it will be ✅ 

```
Using OUTFILE_SUFFIX: .feat
go test -v ./... -run=Test_e2e -count=5 -memprofile=memprofile.feat.out
go: downloading github.com/hashicorp/golang-lru/v2 v2.0.7
go: downloading github.com/Masterminds/semver/v3 v3.2.1
go: downloading go.uber.org/zap v1.27.0
go: downloading github.com/stretchr/testify v1.9.0
go: downloading go.uber.org/multierr v1.11.0
go: downloading github.com/davecgh/go-spew v1.1.1
go: downloading github.com/pmezard/go-difflib v1.0.0
go: downloading gopkg.in/yaml.v3 v3.0.1
go: downloading github.com/stretchr/objx v0.5.2
=== RUN   Test_e2e
=== RUN   Test_e2e/test-case-no-allocations-flag.json
=== RUN   Test_e2e/test-case-no-allocations-flag.json/alice
...
    --- PASS: Test_e2e/test-no-allocations-flag.json (0.00s)
        --- PASS: Test_e2e/test-no-allocations-flag.json/alice (0.00s)
        --- PASS: Test_e2e/test-no-allocations-flag.json/bob (0.00s)
        --- PASS: Test_e2e/test-no-allocations-flag.json/charlie (0.00s)
PASS
ok  	github.com/Eppo-exp/golang-sdk/v5/eppoclient	10.083s
go tool pprof -text memprofile.feat.out > memprofile.feat.text
Switched to a new branch 'main'
branch 'main' set up to track 'origin/main'.
rm -rf eppoclient/test-data/
git clone -b main --depth 1 --single-branch https://github.com/Eppo-exp/sdk-test-data.git eppoclient/test-data/
Cloning into 'eppoclient/test-data'...
Using OUTFILE_SUFFIX: .main
cannot use -memprofile flag with multiple packages
../memprofile.main.out: open ../memprofile.main.out: no such file or directory
failed to fetch any source profiles
make: *** [Makefile:32: profile-memory] Error 1
Failed to generate main profile
total 10568
-rw-r--r-- 1 runner docker     1066 Sep  7 00:17 LICENSE
-rw-r--r-- 1 runner docker     1144 Sep  7 00:18 Makefile
-rw-r--r-- 1 runner docker     7683 Sep  7 00:17 README.md
-rw-r--r-- 1 runner docker       12 Sep  7 00:18 doc.go
drwxr-xr-x 4 runner docker     4096 Sep  7 00:18 eppoclient
-rwxr-xr-x 1 runner docker 10770497 Sep  7 00:18 eppoclient.test
-rw-r--r-- 1 runner docker      454 Sep  7 00:17 go.mod
-rw-r--r-- 1 runner docker     1730 Sep  7 00:17 go.sum
-rw-r--r-- 1 runner docker     4206 Sep  7 00:18 memprofile.feat.out
-rw-r--r-- 1 runner docker     6769 Sep  7 00:18 memprofile.feat.text
-rw-r--r-- 1 runner docker        0 Sep  7 00:18 memprofile.main.text
```